requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
jinja2
numpy
pandas
matplotlib
opencv-python